# govconnect-dashboard/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GovConnect API
  description: |
    Complete API documentation for GovConnect microservices platform.
    
    ## Services
    - **Case Service** - Complaints & Service Requests management
    - **Channel Service** - WhatsApp Channel & Live Chat
    - **AI Service** - AI Orchestrator & Knowledge Base
    - **Notification Service** - Event-driven notifications
    
    ## Authentication
    - **Internal API Key**: `X-Internal-API-Key` header for service-to-service communication
    - **JWT Token**: `Authorization: Bearer <token>` for dashboard authentication
    - **Public APIs**: Endpoint publik tersedia untuk form publik (tanpa auth)
    
    ## Base URLs
    Layanan diakses langsung per service (tanpa gateway agregasi).
  version: 1.0.0
  contact:
    name: GovConnect Team
    url: https://govconnect.my.id
  license:
    name: MIT
servers:
  - url: https://case.govconnect.my.id
    description: Case Service (Production)
  - url: https://channel.govconnect.my.id
    description: Channel Service (Production)
  - url: https://ai.govconnect.my.id
    description: AI Service (Production)
...
components:
  securitySchemes:
    InternalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
      description: Internal API key for service-to-service communication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for dashboard authentication
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
        status:
          type: string
          example: error
...
    ComplaintCategory:
      type: string
      enum:
        - jalan_rusak
        - lampu_mati
        - sampah
        - drainase
        - pohon_tumbang
        - fasilitas_rusak
        - banjir
        - tindakan_kriminal
        - lainnya
    ComplaintStatus:
      type: string
      enum:
        - OPEN
        - PROCESS
        - DONE
        - CANCELED
        - REJECT
    CreateComplaintRequest:
      type: object
      required:
        - wa_user_id
        - kategori
        - deskripsi
      properties:
        wa_user_id:
          type: string
          example: "6281234567890"
        kategori:
          $ref: '#/components/schemas/ComplaintCategory'
        deskripsi:
          type: string
          example: "Jalan berlubang di depan kantor kelurahan"
        alamat:
          type: string
          example: "Jl. Merdeka No. 10"
        rt_rw:
          type: string
          example: "001/002"
        foto_url:
          type: string
          format: uri
          example: "https://example.com/photo.jpg"
    Complaint:
      type: object
      properties:
        id:
          type: string
          format: uuid
        complaint_id:
          type: string
          example: "LP-20251222-001"
        wa_user_id:
          type: string
        kategori:
          $ref: '#/components/schemas/ComplaintCategory'
        deskripsi:
          type: string
        alamat:
          type: string
        rt_rw:
          type: string
        foto_url:
          type: string
        status:
          $ref: '#/components/schemas/ComplaintStatus'
        admin_notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /laporan/create:
    post:
      tags:
        - Complaints
      summary: Create complaint
      description: Create a new complaint/laporan
      operationId: createComplaint
      security:
        - InternalApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComplaintRequest'
      responses:
        '201':
          description: Complaint created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Complaint'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /laporan/{id}:
    get:
      tags:
        - Complaints
      summary: Get complaint by ID
      operationId: getComplaintById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Complaint details
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    $ref: '#/components/schemas/Complaint'
        '404':
          description: Complaint not found
...
  /laporan/{id}/status:
    patch:
      tags:
        - Complaints
      summary: Update complaint status
      operationId: updateComplaintStatus
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: '#/components/schemas/ComplaintStatus'
                admin_notes:
                  type: string
      responses:
        '200':
          description: Status updated
        '404':
          description: Complaint not found
  /laporan/{id}/cancel:
    post:
      tags:
        - Complaints
      summary: Cancel complaint
      operationId: cancelComplaint
      security:
        - InternalApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - wa_user_id
              properties:
                wa_user_id:
                  type: string
                cancel_reason:
                  type: string
      responses:
        '200':
          description: Complaint cancelled
        '403':
          description: Not owner
        '404':
          description: Not found


  # ============================================
  # CASE SERVICE - Statistics & User
  # ============================================
...
  /user/{wa_user_id}/history:
    get:
      tags:
        - User
      summary: Get user history
      description: Get user history including complaints and service requests
      operationId: getUserHistory
      security:
        - InternalApiKey: []
      parameters:
        - name: wa_user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User history


  # ============================================
  # CHANNEL SERVICE - Webhook
  # ============================================
...